+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Gluecodium project Release Notes
 
+## Unreleased
+### Bug fixes:
+  * Fixed compilation issues in Swift and Dart when inheriting from an interface with overloaded functions.
+
 ## 11.1.7
 Release date: 2022-05-05
 ### Bug fixes: